Files were in a subfolder instead of the working directory
