--- Systemdesign-Projekt/data/en/ui.json
@@ -1,9 +1,0 @@
-{
-  "ordersFinished" : "Finshed orders",
-  "ordersInQueue" : "Order queue",
-  "placeOrder" : "Place order",
-  "mains" : "Main courses",
-  "sideorders" : "Side orders",
-  "drinks" : "Drinks",
-  "ready" : "Done"
-}
