--- Systemdesign-Projekt/data/sv/menu.json
@@ -1,191 +1,0 @@
-{
-  "foods": [
-    {
-      "order": 1,
-      "label": "Hamburgare",
-      "price": 65,
-      "details": [
-        { "extra": "Lök",
-          "price": 0
-        }
-      ]
-    },
-    {
-      "order": 2,
-      "label": "Ostburgare",
-      "price": 70,
-      "details": [
-        { "extra": "Lök",
-          "price": 0
-        },
-        { "extra": "Ost",
-          "price": 10
-        }
-      ]
-    },
-    {
-      "order": 3,
-      "label": "Veganburgare",
-      "price": 65,
-      "details": [
-        { "extra": "Lök",
-          "price": 0
-        },
-        { "extra": "Tryffel",
-          "price": 20
-        },
-        { "extra": "Bacon",
-          "price": 10
-        }
-      ]
-    }
-  ],
-  "drinks": [
-  {
-    "order": 1,
-    "label": "Fullers India Pale Ale",
-    "size": 50,
-    "price": 65
-  },
-  {
-    "order": 2,
-    "label": "Fullers ESB",
-    "size": 50,
-    "price": 65
-  },
-  {
-    "order": 3,
-    "label": "Brew Dog 5 a.m",
-    "size": 33,
-    "price": 65
-  },
-  {
-    "order": 4,
-    "label": "Duvel",
-    "size": 33,
-    "price": 65
-  },
-  {
-    "order": 5,
-    "label": "Wisby Oak Porter",
-    "size": 33,
-    "price": 65
-  },
-  {
-    "order": 6,
-    "label": "Wisby Klosteröl",
-    "size": 33,
-    "price": 65
-  },
-  {
-    "order": 7,
-    "label": "Baron Trenck",
-    "size": 50,
-    "price": 65
-  },
-  {
-    "order": 8,
-    "label": "Adnams Bitter",
-    "size": 50,
-    "price": 65
-  },
-  {
-    "order": 9,
-    "label": "Innis & Gunn Orginal",
-    "size": 33,
-    "price": 65
-  },
-  {
-    "order": 10,
-    "label": "Innis & Gunn Toasted Oak IPA",
-    "size": 33,
-    "price": 65
-  },
-  {
-    "order": 11,
-    "label": "Liberty Ale",
-    "size": 33,
-    "price": 65
-  },
-  {
-    "order": 12,
-    "label": "Anchor Steam",
-    "size": 33,
-    "price": 65
-  },
-  {
-    "order": 13,
-    "label": "Nanny State",
-    "size": 33,
-    "price": 65,
-    "details": [
-      {
-        "non-alcoholic":true
-      }
-    ]
-  },
-  {
-    "order": 14,
-    "label": "Störtebeker",
-    "size": 33,
-    "price": 65,
-    "details": [
-      {
-        "non-alcoholic":true
-      }
-    ]
-  },
-  {
-    "order": 15,
-    "label": "Zatec",
-    "size": 33,
-    "price": 65,
-    "details": [
-      {
-        "non-alcoholic":true,
-        "gluten-free":true
-      }
-    ]
-  },
-  {
-    "order": 16,
-    "label": "Green’s Blond Ale",
-    "size": 33,
-    "price": 50,
-    "details": [
-      {
-        "non-alcoholic":true,
-        "gluten-free":true
-      }
-    ]
-  }
-],
-  "sideorders": [
-    {
-      "order": 1,
-      "label": "Strips",
-      "price": 20,
-      "details": [
-        { "extra": "Smält ost",
-          "price": 10
-        },
-        { "extra": "Stekt lök",
-          "price": 10
-        }
-      ]
-  },
-      {
-      "order": 1,
-      "label": "Nachos",
-      "price": 30,
-      "details": [
-        { "extra": "Smält ost",
-          "price": 10
-        },
-        { "extra": "Salsa",
-          "price": 10
-        }
-      ]
-  }
-]
-}
