--- Systemdesign-Projekt/data/sv/ui.json
@@ -1,9 +1,0 @@
-{
-  "ordersFinished" : "Färdiga beställningar",
-  "ordersInQueue" : "Beställningar att förbereda",
-  "placeOrder" : "Lägg beställning",
-  "mains" : "Huvudrätter",
-  "sideorders" : "Tillbehör",
-  "drinks" : "Drycker",
-  "ready" : "Klart"
-}
