--- Systemdesign-Projekt/public/css/nationPub.css
@@ -1,83 +1,0 @@
-@import 'https://fonts.googleapis.com/css?family=Open+Sans';
-
-header{
-	margin: 10px 10px 10px 10px;
-}
-body{
-	font-family: 'Open Sans', sans-serif;
-	font-size: 20px;
-	margin: 10px 10px 10px 10px;
-}
-h1{
-	font-size: 52px;
-}
-#title{
-	color: black;
-	position: absolute;
-	margin: 30px 10px 10px 10px;
-	z-index: 0;
-}
-#crop{
-	height: 100px;
-	overflow: hidden;
-	opacity: 0.8;
-	z-index: 1;
-}
-button{
-	background-color: white;
-    border: none;
-    color: white;
-}
-.test{
-	table-layout: fixed;
-	height: 550px;
-	overflow-y: scroll;
-	float: left;
-
-}
-table#drink{
-	padding-right: 10px;
-	border-right: 2px solid black;
-}
-table#food{
-	padding-right: 10px;
-	border-right: 2px solid black;
-}
-.amount{
-	table-layout: fixed;
-	height: 500px;
-	float: left;
-	margin: 0px 5px 5px 10px
-}
-.bord{
-	table-layout: fixed;
-	height: 500px;
-	float: left;
-	margin: -2px 5px 5px 10px
-}
-.bord1{
-	table-layout: fixed;
-	height: 500px;
-	float: left;
-	margin: 0px 5px 5px 5px
-}
-.sendorder{
-	margin: 0px 5px 5px 5px;
-	table-layout: fixed;
-	height: 50px;
-	float: left;
-}
-.price{
-	margin: 0px 5px 5px 5px;
-	table-layout: fixed;
-	height: 50px;
-	float: left;
-}
-.removeItem{
-	margin: 0px 5px 5px 5px;
-	table-layout: fixed;
-	border: 2px solid black;
-	height: 50px;
-	float: left;
-}
-
