--- Systemdesign-Projekt/views/diner.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head id="head">
-    <title>Wermlandskällar'n</title>
-    <meta charset="utf-8"/>
-    <link rel="stylesheet" type="text/css" href="css/nationPub.css">
-    <script type="text/javascript" src="data/en/menu.json"></script>
-    <script type="text/javascript" src="js/nationPub.js"></script>
-    <script type="text/javascript" src="js/diner.js"></script>
-    <script>
-      docLoaded(indexPageLoaded);
-    </script>
-  </head>
-  <body>
-    <main>
-      <header>
-      <h1><span id="title">Wermlandskällar'n</h1>
-      <div id="crop"><img src="https://s-media-cache-ak0.pinimg.com/originals/47/75/5c/47755cd9af91491f1f919afae3ce8a0a.jpg"></div>
-      </header>
-      <section class="test"><table id="drink"></table></section>
-      <section class="test"><table id="food"></table></section>
-      <div class="removeItem"><section class="remove"><table id="remove"></table></section></div>
-      <div class="amount"><section class="test1"><table id="x"></table></section></div>
-      <div class="bord"><section class="test1"><table id="order"></table></section></div>
-      <div class="bord1"><section class="test1"><table id="price"></table></section></div>
-      <div class="sendorder"><section class="send"><table id="send"></table></section></div>
-      <div class="price"><section class="total"><table id="total"></table></section></div>
-    </main>
-  </body>
-</html>
-
-
