--- Systemdesign-Projekt/views/kitchen.html
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Nationsköket</title>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" href="/css/kitchen.css">
-  <script src="/js/kitchenScript.js"></script>
-  <script src="https://unpkg.com/vue/dist/vue.js"></script>
-</head>
-<body>
-    <header>
-        <script>docLoaded(kitchenPageLoaded);</script>
-        KITCHEN
-    </header>
-
-    <section>
-        <button v-on:click='send' id='knapp' type=submit></button>
-
-        <table>
-            <thead>
-                <tr>
-                    <th>Ongoing orders</th>
-                    <th>Finished orders</th>
-                </tr>
-            </thead>
-            <tbody>
-                <tr>
-                    <td id='ongoing'>  </td>
-                    <td id='completed'></td>
-                </tr>
-            </tbody>
-        </table>
-
-    </section>
-    <script src="/js/kitchenVue.js"></script>
-</body>
-</html>
-
-<!--
-<body>
-  <div id="orders">
-    <h1>{{ uiLabels.ordersInQueue }}</h1>
-    <ol>
-      <li v-for="(item, key) in orders" v-if="!item.done">
-        {{ key }} : {{ item.orderItems.join(', ') }}
-        <button v-on:click="markDone(key)">
-          {{uiLabels.ready}}
-        </button>
-      </li>
-    </ol>
-        <h1>{{ uiLabels.ordersFinished }}</h1>
-    <ol>
-      <li v-for="(item, key) in orders" v-if="item.done">
-        {{ key }} : {{ item.orderItems.join(', ') }}
-      </li>
-    </ol>
-  </div>
-  <script src="/vue/vue.js"></script>
-  <script src="/socket.io/socket.io.js"></script>
-  <script src="js/nationPub.js"></script>
-  <script src="js/kitchen.js"></script>
-</body>
-</html>
--->
