+++ views/diner.html
@@ -18,9 +18,13 @@
       <div id="crop"><img src="https://s-media-cache-ak0.pinimg.com/originals/47/75/5c/47755cd9af91491f1f919afae3ce8a0a.jpg"></div>
       </header>
       <section class="test"><table id="drink"></table></section>
-      <section class="test"><table id="food"></table></section> 
-      <div class="bord"><section class="test"><table id="order"></table></section></div>
-      <div class="bord1"><section class="test"><table id="price"></table></section></div>
+      <section class="test"><table id="food"></table></section>
+      <div class="removeItem"><section class="remove"><table id="remove"></table></section></div>
+      <div class="amount"><section class="test1"><table id="x"></table></section></div>
+      <div class="bord"><section class="test1"><table id="order"></table></section></div>
+      <div class="bord1"><section class="test1"><table id="price"></table></section></div>
+      <div class="sendorder"><section class="send"><table id="send"></table></section></div>
+      <div class="price"><section class="total"><table id="total"></table></section></div>
     </main>
   </body>
 </html>
